添加了PIT

+++ Sources/supersonic.c
@@ -11,7 +11,7 @@ int distance=0;
 
 void intc_get_supersonic_time_0(void)
 {
-	unsigned long tmp_a, tmp_b;
+	int tmp_a, tmp_b;
 	tmp_a = EMIOS_0.CH[7].CADR.R;
 	tmp_b = EMIOS_0.CH[7].CBDR.R;
 	if(tmp_a >= tmp_b)
